backend/requirements.txt
fastapi
uvicorn[standard]
SQLAlchemy
psycopg2-binary
python-dotenv
pydantic-settings
email-validator
python-jose[cryptography]
python-multipart
passlib
bcrypt==4.3.0
